# Don't allocate memory for png decode filters, reuse scanline buffers