… tests/ImageSharp.Tests/TestImages/Formats/Png/filter0.png
… tests/ImageSharp.Tests/TestImages/Formats/Png/filter1.png
… tests/ImageSharp.Tests/TestImages/Formats/Png/filter2.png
… tests/ImageSharp.Tests/TestImages/Formats/Png/filter3.png
… tests/ImageSharp.Tests/TestImages/Formats/Png/filter4.png
… tests/ImageSharp.Tests/TestImages/Formats/Png/filterVar.png