This PR is part of perf improvements
- Stop allocating memory for png decode filters, reuse scanline buffers
- Added png test images with different filters for local tests

P.S. This passed on test data before your last encoder commit, need to retest once you have fix...
P.S2 This revealed a bug in decoding 4bbp png pictures (the last one in this PR), decoder just crashes, needs to be fixed separately