@JimBobSquarePants 2016-11-11
Thanks for this! Good to see and really appreciate it. 

Do you have any benchmarks for before and after?

I'll get a fix for the encoder asap. Will have to roll back to a working version and incrementally make changes to figure out what is wrong.

The 4bpp fix should be easy enough I'm sure.