@boguscoder 2016-11-11
Oh yeah, I'll provide benchmarks here soon