@boguscoder 2016-11-11
I am actually not happy with perf results, let me dig a bit deeper