@boguscoder 2016-11-11
Ok, Im seeing really nano improvements here:

Before:

                    Method |      Median |    StdDev | Scaled | Scaled-SD |
-------------------------- |------------ |---------- |------- |---------- |
    None-filtered PNG file |  74.1752 us | 1.4026 us |   1.00 |      0.00 |
     Sub-filtered PNG file |  59.0522 us | 1.3691 us |   0.79 |      0.02 |
      Up-filtered PNG file |  68.8289 us | 2.6562 us |   0.93 |      0.04 |
 Average-filtered PNG file |  91.3585 us | 1.1904 us |   1.23 |      0.03 |
   Paeth-filtered PNG file | 108.2761 us | 6.1980 us |   1.44 |      0.09 |

After:

                    Method |      Median |     StdDev | Scaled | Scaled-SD |
-------------------------- |------------ |----------- |------- |---------- |
    None-filtered PNG file |  73.1292 us |  1.9591 us |   1.00 |      0.00 |
     Sub-filtered PNG file |  58.1899 us |  2.8178 us |   0.79 |      0.04 |
      Up-filtered PNG file |  65.5541 us |  3.1199 us |   0.90 |      0.05 |
 Average-filtered PNG file |  58.2450 us |  3.2539 us |   0.81 |      0.05 |
   Paeth-filtered PNG file | 106.6377 us | 17.3648 us |   1.33 |      0.23 |


Note that your changes added to 'after' case and Im not sure that adding 'fixed' approach to filters really made it faster, I think I got some intermediate test results that showed slow down