@dlemstra 2016-11-11
There should be less allocations though because you are using the input array. Still think it should be good to merge your changes.