@boguscoder 2016-11-11
Even if I perform tests on the very same file I get rather different results, Im not sure if its because its that fast and timing it is hard or some other reason, so I would not trust numbers above a lot, but in general my changes should make sense..