@JimBobSquarePants 2016-11-12
@boguscoder The improvements were more about memory usage. I'll merge this and run some tests locally regarding the fixed keyword usage.  I'll have a look then at fixing the  4bbp png issue now that I have fixed the encoder.