@JimBobSquarePants 2016-11-12
@boguscoder  Here's the current state of play with the decoder with your changes and a couple of additional tweaks.

```
Host Process Environment Information:
BenchmarkDotNet-Dev.Core=v0.9.9.0
OS=Microsoft Windows NT 6.2.9200.0
Processor=Intel(R) Core(TM) i7-6700HQ CPU 2.60GHz, ProcessorCount=8
Frequency=2531249 ticks, Resolution=395.0619 ns, Timer=TSC
CLR=MS.NET 4.0.30319.42000, Arch=64-bit RELEASE [RyuJIT]
GC=Concurrent Workstation
JitModules=clrjit-v4.6.1586.0

Type=DecodePng  Mode=Throughput
```
             Method |     Median |    StdDev | Scaled | Scaled-SD | Gen 0 | Gen 1 |  Gen 2 | Bytes Allocated/Op |
------------------- |----------- |---------- |------- |---------- |------ |------ |------- |------------------- |
 System.Drawing Png |  3.1796 ms | 0.1052 ms |   1.00 |      0.00 |     - |     - |  47.10 |          90,395.43 |
     ImageSharp Png | 11.7749 ms | 0.4999 ms |   3.72 |      0.19 | 69.00 |  1.00 | 178.00 |         503,921.20 |

That's half the memory we were using before so a nice improvement. 

The obvious memory hog right now is the allocation of memory when assigning chunk data. Any additional performance improvements I'm not sure about.